+++ spec/config/memoized_value_spec.rb
@@ -1,0 +1,9 @@
+RSpec.describe Pry::Config::MemoizedValue do
+  describe "#call" do
+    subject { described_class.new { rand } }
+
+    it "memoizes the result of call" do
+      expect(subject.call).to eq(subject.call)
+    end
+  end
+end